docker_requirements.txt
# Base data science and ML packages
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
scipy>=1.11.0
matplotlib>=3.8.0
seaborn>=0.12.0

# ML specific packages
xgboost>=2.1.0
catboost>=1.2.0
shap>=0.48.0

# Kedro and MLflow
kedro>=0.19.0
kedro-mlflow>=0.12.0
kedro-viz>=9.0.0
mlflow>=2.11.0

# Data validation
great_expectations>=0.18.0
ruamel.yaml>=0.17.17
hsfs>=3.7.0  # Hopsworks Feature Store - required for Great Expectations

# Visualization
streamlit>=1.32.0
plotly>=5.21.0

# Other essential packages
pyyaml>=6.0
scikit-optimize>=0.9.0
pytest>=7.0.0
tqdm>=4.66.0
fsspec>=2024.3.0
pyarrow>=15.0.0
pymysql>=1.0.0  # Often required by HSFS
requests>=2.28.0  # Required for API interactions

# Dependencies that are likely causing issues - use specific versions
jinja2==3.1.3  # Often has compatibility issues
markupsafe==2.1.5  # Often has compatibility issues
